Rename assets-manager
