+++ src/core/assets-manager.ts
